backend/requirements.txt
# NotesOS Backend

## Requirements
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==3.2.2
python-multipart>=0.0.6
httpx>=0.26.0
redis>=5.0.1
pgvector>=0.2.4
pydantic[email]
psycopg2-binary==2.9.11
# AI & ML
langchain>=0.1.0
langgraph>=0.0.20
anthropic>=0.18.0
openai>=1.10.0
voyageai>=0.2.0

# Utilities
python-magic>=0.4.27
beautifulsoup4>=4.12.0
pdf2image>=1.16.3
pytesseract>=0.3.10
mammoth>=1.6.0
Pillow>=10.0.0

# File Storage
cloudinary>=1.36.0

# Google Cloud (optional - for premium OCR fallback)
google-cloud-vision>=3.4.0
